+++ documentation/sidebars.js
@@ -394,6 +394,7 @@ module.exports = {
             type: "category",
             label: "Packages",
             items: [
+                "packages/command-palette",
                 "packages/react-table",
                 {
                     type: "category",
@@ -515,6 +516,7 @@ module.exports = {
                     ],
                 },
                 "examples/antd-calendar-example",
+                "examples/command-palette",
                 {
                     type: "category",
                     label: "Core",