+++ examples/base/antd/src/pages/posts/list.tsx
@@ -15,9 +15,9 @@ import {
     Select,
     Radio,
     TagField,
+    useTable,
+    useSelect,
 } from "@pankod/refine-antd";
-
-import { useTable, useSelect } from "@pankod/refine-antd";
 
 import { IPost, ICategory } from "interfaces";