+++ examples/tutorial/mui/src/pages/posts/list.tsx
@@ -95,7 +95,8 @@ export const PostList: React.FC = () => {
                 },
             },
         ],
-        [],
+        // highlight-next-line
+        [categoriesData, isLoading],
     );
 
     return (