+++ packages/antd/src/hooks/modal/useModal/index.tsx
@@ -21,6 +21,7 @@ export const useModal = ({
 
     return {
         modalProps: {
+            ...modalProps,
             onCancel: (e: React.MouseEvent<HTMLElement>) => {
                 modalProps.onCancel?.(e);
                 close();