+++ packages/core/src/hooks/data/useList.ts
@@ -7,8 +7,8 @@ import {
     HttpError,
     CrudSorting,
     MetaDataQuery,
+    SuccessErrorNotification,
     LiveModeProps,
-    OpenNotificationParams,
 } from "../../interfaces";
 import {
     useCheckError,
@@ -30,11 +30,10 @@ export type UseListProps<TData, TError> = {
     resource: string;
     config?: UseListConfig;
     queryOptions?: UseQueryOptions<GetListResponse<TData>, TError>;
-    successNotification?: OpenNotificationParams | false;
-    errorNotification?: OpenNotificationParams | false;
     metaData?: MetaDataQuery;
     dataProviderName?: string;
-} & LiveModeProps;
+} & SuccessErrorNotification &
+    LiveModeProps;
 
 /**
  * `useList` is a modified version of `react-query`'s {@link https://react-query.tanstack.com/guides/queries `useQuery`} used for retrieving items from a `resource` with pagination, sort, and filter configurations.
@@ -109,13 +108,28 @@ export const useList = <
             ...queryOptions,
             onSuccess: (data) => {
                 queryOptions?.onSuccess?.(data);
-                handleNotification(successNotification);
+
+                const notificationConfig =
+                    typeof successNotification === "function"
+                        ? successNotification(
+                              data,
+                              { metaData, config },
+                              resource,
+                          )
+                        : successNotification;
+
+                handleNotification(notificationConfig);
             },
             onError: (err: TError) => {
                 checkError(err);
                 queryOptions?.onError?.(err);
 
-                handleNotification(errorNotification, {
+                const notificationConfig =
+                    typeof errorNotification === "function"
+                        ? errorNotification(err, { metaData, config }, resource)
+                        : errorNotification;
+
+                handleNotification(notificationConfig, {
                     key: `${resource}-useList-notification`,
                     message: translate(
                         "common:notifications.error",