+++ packages/nhost/src/dataProvider/index.ts
@@ -81,7 +81,9 @@ export const generateFilters: any = (filters?: CrudFilters) => {
         }
 
         if (filter.operator !== "or") {
-            resultFilter[filter.field] = {};
+            if (!resultFilter.hasOwnProperty(filter.field)) {
+                resultFilter[filter.field] = {};
+            }
             resultFilter[filter.field][operator] = filter.value;
         } else {
             const orFilter: any = [];
@@ -95,7 +97,9 @@ export const generateFilters: any = (filters?: CrudFilters) => {
                         `Operator ${val.operator} is not supported`,
                     );
                 }
-                filterObject[val.field] = {};
+                if (!filterObject.hasOwnProperty(val.field)) {
+                    filterObject[val.field] = {};
+                }
                 filterObject[val.field][mapedOperator] = val.value;
                 orFilter.push(filterObject);
             });